Consolidate test cases and fixes
